requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.4.0
python-multipart==0.0.9

pymupdf==1.24.11
docling==2.80.0

colpali-engine==0.3.10
torch==2.5.1
torchvision==0.20.1
Pillow==10.4.0

qdrant-client==1.11.3

openai==1.51.2

python-dotenv==1.0.1
httpx==0.27.2
numpy==1.26.4
tqdm==4.66.5
duckdb==1.2.1
aiofiles==23.2.1
